Moved Project sync to its own component Now it is at the project root level, so it can be invoked no matter what screen the user is on
