+++ common/src/commonMain/kotlin/com/darkrockstudios/apps/hammer/common/components/projecthome/ProjectHome.kt
@@ -1,7 +1,6 @@
 package com.darkrockstudios.apps.hammer.common.components.projecthome
 
 import com.arkivanov.decompose.value.Value
-import com.darkrockstudios.apps.hammer.base.http.ApiProjectEntity
 import com.darkrockstudios.apps.hammer.common.components.projectroot.Router
 import com.darkrockstudios.apps.hammer.common.data.ProjectDef
 import com.darkrockstudios.apps.hammer.common.data.encyclopediarepository.entry.EntryType
@@ -13,10 +12,7 @@ interface ProjectHome : Router, HammerComponent {
 	suspend fun exportProject(path: String)
 	fun beginProjectExport()
 	fun endProjectExport()
-	fun syncProject()
-	fun resolveConflict(resolvedEntity: ApiProjectEntity)
-	fun endSync()
-	fun cancelSync()
+	fun startProjectSync()
 
 	data class State(
 		val projectDef: ProjectDef,
@@ -27,39 +23,5 @@ interface ProjectHome : Router, HammerComponent {
 		val encyclopediaEntriesByType: Map<EntryType, Int> = emptyMap(),
 		val showExportDialog: Boolean = false,
 		val hasServer: Boolean = false,
-		val isSyncing: Boolean = false,
-		val syncProgress: Float = 0f,
-		val entityConflict: EntityConflict<*>? = null,
-		val syncLog: List<String> = emptyList(),
 	)
-
-	sealed class EntityConflict<T : ApiProjectEntity>(
-		val serverEntity: T,
-		val clientEntity: T
-	) {
-		class SceneConflict(
-			serverScene: ApiProjectEntity.SceneEntity,
-			clientScene: ApiProjectEntity.SceneEntity
-		) : EntityConflict<ApiProjectEntity.SceneEntity>(serverScene, clientScene)
-
-		class NoteConflict(
-			serverNote: ApiProjectEntity.NoteEntity,
-			clientNote: ApiProjectEntity.NoteEntity
-		) : EntityConflict<ApiProjectEntity.NoteEntity>(serverNote, clientNote)
-
-		class TimelineEventConflict(
-			serverEvent: ApiProjectEntity.TimelineEventEntity,
-			clientEvent: ApiProjectEntity.TimelineEventEntity
-		) : EntityConflict<ApiProjectEntity.TimelineEventEntity>(serverEvent, clientEvent)
-
-		class EncyclopediaEntryConflict(
-			serverEntry: ApiProjectEntity.EncyclopediaEntryEntity,
-			clientEntry: ApiProjectEntity.EncyclopediaEntryEntity
-		) : EntityConflict<ApiProjectEntity.EncyclopediaEntryEntity>(serverEntry, clientEntry)
-
-		class SceneDraftConflict(
-			serverEntry: ApiProjectEntity.SceneDraftEntity,
-			clientEntry: ApiProjectEntity.SceneDraftEntity
-		) : EntityConflict<ApiProjectEntity.SceneDraftEntity>(serverEntry, clientEntry)
-	}
 }
