+++ composeUi/src/commonMain/kotlin/com/darkrockstudios/apps/hammer/common/projectsync/ConflictResolution.kt
@@ -1,86 +1,21 @@
-package com.darkrockstudios.apps.hammer.common.projecthome
+package com.darkrockstudios.apps.hammer.common.projectsync
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material3.Button
-import androidx.compose.material3.LinearProgressIndicator
 import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
-import com.darkrockstudios.apps.hammer.common.components.projecthome.ProjectHome
-import com.darkrockstudios.apps.hammer.common.compose.MpDialog
+import com.darkrockstudios.apps.hammer.common.components.projectsync.ProjectSync
 import com.darkrockstudios.apps.hammer.common.compose.Ui
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
 @Composable
-internal fun ProjectSynchronization(state: ProjectHome.State, component: ProjectHome) {
-	MpDialog(
-		title = "Project Synchronization",
-		onCloseRequest = {},
-		visible = state.isSyncing,
-	) {
-		val scope = rememberCoroutineScope()
-
-		Column(modifier = Modifier.padding(Ui.Padding.L).fillMaxSize()) {
-			Text("Project Synchronizing...")
-			Spacer(modifier = Modifier.size(Ui.Padding.L))
-			LinearProgressIndicator(
-				progress = state.syncProgress,
-				modifier = Modifier.fillMaxWidth()
-			)
-			if (state.syncProgress == 1f) {
-				Button(onClick = component::endSync) {
-					Text("Complete")
-				}
-			} else {
-				Button(onClick = component::cancelSync) {
-					Text("Cancel")
-				}
-			}
-			val conflict = state.entityConflict
-			if (conflict != null) {
-				when (conflict) {
-					is ProjectHome.EntityConflict.SceneConflict -> {
-						val sceneConflict = state.entityConflict as ProjectHome.EntityConflict.SceneConflict
-						SceneConflict(sceneConflict, component)
-					}
-
-					is ProjectHome.EntityConflict.NoteConflict -> {
-						val noteConflict = state.entityConflict as ProjectHome.EntityConflict.NoteConflict
-						NoteConflict(noteConflict, component)
-					}
-
-					is ProjectHome.EntityConflict.TimelineEventConflict -> {
-						val timelineEventConflict =
-							state.entityConflict as ProjectHome.EntityConflict.TimelineEventConflict
-						TimelineEventConflict(timelineEventConflict, component)
-					}
-
-					is ProjectHome.EntityConflict.EncyclopediaEntryConflict -> {
-						val encyclopediaEntryConflict =
-							state.entityConflict as ProjectHome.EntityConflict.EncyclopediaEntryConflict
-						EncyclopediaEntryConflict(encyclopediaEntryConflict, component)
-					}
-
-					is ProjectHome.EntityConflict.SceneDraftConflict -> {
-						val sceneDraftConflict =
-							state.entityConflict as ProjectHome.EntityConflict.SceneDraftConflict
-						SceneDraftConflict(sceneDraftConflict, component)
-					}
-				}
-			} else {
-				SyncLog(state, scope)
-			}
-		}
-	}
-}
-
-@Composable
-internal fun SyncLog(state: ProjectHome.State, scope: CoroutineScope) {
+internal fun SyncLog(state: ProjectSync.State, scope: CoroutineScope) {
 	var showLog by rememberSaveable { mutableStateOf(true) }
 	Column(modifier = Modifier.fillMaxWidth()) {
 		Button(onClick = { showLog = !showLog }) {
@@ -107,8 +42,8 @@ internal fun SyncLog(state: ProjectHome.State, scope: CoroutineScope) {
 
 @Composable
 internal fun SceneConflict(
-	entityConflict: ProjectHome.EntityConflict.SceneConflict,
-	component: ProjectHome
+	entityConflict: ProjectSync.EntityConflict.SceneConflict,
+	component: ProjectSync
 ) {
 	Column {
 		Text("Scene Conflict")
@@ -135,8 +70,8 @@ internal fun SceneConflict(
 
 @Composable
 internal fun NoteConflict(
-	entityConflict: ProjectHome.EntityConflict.NoteConflict,
-	component: ProjectHome
+	entityConflict: ProjectSync.EntityConflict.NoteConflict,
+	component: ProjectSync
 ) {
 	Column {
 		Text("Note Conflict")
@@ -163,8 +98,8 @@ internal fun NoteConflict(
 
 @Composable
 internal fun TimelineEventConflict(
-	entityConflict: ProjectHome.EntityConflict.TimelineEventConflict,
-	component: ProjectHome
+	entityConflict: ProjectSync.EntityConflict.TimelineEventConflict,
+	component: ProjectSync
 ) {
 	Column {
 		Text("Timeline Event Conflict")
@@ -191,8 +126,8 @@ internal fun TimelineEventConflict(
 
 @Composable
 internal fun EncyclopediaEntryConflict(
-	entityConflict: ProjectHome.EntityConflict.EncyclopediaEntryConflict,
-	component: ProjectHome
+	entityConflict: ProjectSync.EntityConflict.EncyclopediaEntryConflict,
+	component: ProjectSync
 ) {
 	Column {
 		Text("Encyclopedia Entry Conflict")
@@ -219,8 +154,8 @@ internal fun EncyclopediaEntryConflict(
 
 @Composable
 internal fun SceneDraftConflict(
-	entityConflict: ProjectHome.EntityConflict.SceneDraftConflict,
-	component: ProjectHome
+	entityConflict: ProjectSync.EntityConflict.SceneDraftConflict,
+	component: ProjectSync
 ) {
 	Column {
 		Text("Scene Draft Conflict")
